+++ app/google/oauth/index.tsx
@@ -4,9 +4,8 @@ import { useSetAtom } from 'jotai';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect } from 'react';
 
-import { LoadingArea } from '@/components/atoms';
+import { LoginLoading, LoginScreen } from '@/features/login';
 import { AuthInfoAtom } from '@/store/auth';
-import { flex } from '@/styled-system/patterns';
 import { AuthResponse } from '@/types/authType';
 
 const Page = () => {
@@ -57,20 +56,11 @@ const Page = () => {
   }, [router, searchParams, setAuth]);
 
   return (
-    <div className={LoadingWrapper}>
-      <LoadingArea width={100} height={100} />
-      <div>구글로 로그인중입니다.</div>
-    </div>
+    <>
+      <LoginLoading />
+      <LoginScreen isAnimate={false} />
+    </>
   );
 };
-
-const LoadingWrapper = flex({
-  direction: 'column',
-  justifyContent: 'center',
-  alignContent: 'center',
-  height: '100vh',
-  textAlign: 'center',
-  fontWeight: 'bold',
-});
 
 export default Page;
