+++ components/atoms/loading/index.ts
@@ -1,1 +1,2 @@
 export * from './loading';
+export * from './loading-spinner';
